+++ R/RcppExports.R
@@ -5,3 +5,7 @@ matMultinom <- function(probmatrix) {
     .Call(`_simstudy_matMultinom`, probmatrix)
 }
 
+markovChains <- function(nchains, P, chainLen) {
+    .Call(`_simstudy_markovChains`, nchains, P, chainLen)
+}
+